Enable source stepping to work with transient only source
The source stepping homotopy method can only scale the default
parameter of the sources which is a DC source data parameter. In
order to make source stepping work with sources that only have
transient specification with nonzero initial values, the original value of the default
parameter needs to be overwritten so the inital value in the
transient source can be scaled during the source stepping. It needs
to be done only once before source stepping starts. This commit
enables the source stepping method to work with transient only sources.
This improves the robustness and efficiency of the source stepping
for the circuits that have transient only source with nonzero initial value

https://cee-gitlab.sandia.gov/Xyce/backlog/project-backlog/-/issues/622