Fix adjacency count in parallel for sequential source stepping
While generating a parallel test for sequential source stepping, it
was noticed that the adjacency count was not correct in parallel.
Off processor device connectivity was not being captured correctly.
This resulted in the sequential source stepping being less effective
as the number of processors increased.  This has been fixed now.

https://cee-gitlab.sandia.gov/Xyce/backlog/project-backlog/-/issues/634