Enable source stepping to work with sources with inconsistent DC and transient specification
The commit fixes the voltage source so source stepping homotopy
method can work with cases that have inconsistent DC and transient
specifications. When a voltage source specifies both DC and transient
values, it is possible that the DC value is inconsistent to the
transient specification at time 0. Which value to use during source
stepping depends on the analysis. It also needs to be consistent
to the case when standard newton method (noncontinuation method) is
used for DC op calculation. These are fixed for the voltage source.
This commit also enables source stepping to work with AC analysis for
different cases. The AC analysis solve DC op first before the AC
calculation. The fixes improve robustness of source stepping method

https://cee-gitlab.sandia.gov/Xyce/backlog/project-backlog/-/issues/622