Enable source stepping to work with current source with transient specification
The source stepping can work with current sources that only have DC
specification previously. The fix on issue 583 enables source stepping to
work with the case when both DC and transient values are specified.
But these values need to be consistent. This commit enables
source stepping to work with other cases that involve transient
specification in the current source, such as transient only,
inconsistent DC and transient values. The value that is scaled
during source stepping will be overwritten in some cases and this
depends on the analysis type, solver status, different combination
of source type specified in the current source

https://cee-gitlab.sandia.gov/Xyce/backlog/project-backlog/-/issues/622