Fix the device instance to be consistent to device master
The voltage and current sources have device loading functions
for both device instance and master classes. The recent improvement
on source stepping need to overwrite the original value of the
default parameter and it needs to happen only once. Normally, the
first time device loading happens in device master. And the code
change related to source stepping is in the master class. This commit
fixes the device instance class so it is consitent to the master class.

This commit fixes both voltage and current sources

https://cee-gitlab.sandia.gov/Xyce/backlog/project-backlog/-/issues/622