Update the users' guide about improvement on source stepping
This commit updated users' guide about improvement on source
stepping. It now works for all sources, not just DC only sources.
Note any, or all DC, AC and transient values can be specified in
sources. The document is also added for different
cases when inconsistent values are specified and which value is
used during source stepping in the users' guide

https://cee-gitlab.sandia.gov/Xyce/backlog/project-backlog/-/issues/622