Fix global param dependency in subckt params
This commit contains fixes for both of these issues:

https://cee-gitlab.sandia.gov/Xyce/backlog/project-backlog/-/issues/631
"Expression based local variation doesn't always work with subcircuit
arguments"

https://cee-gitlab.sandia.gov/Xyce/backlog/project-backlog/-/issues/609
"Global parameters that are applied to subcircuit parameters don't work
in parallel"

For both issues, the term "subcircuit parameters" refers to parameters
found on the "X" line of a netlist.  In other words, params used in the
invocation of a netlist, not the netlist definition (which would be a
.SUBCKT line).

The problem for both of these issues had to do with the order in which
parameters were resolved, and when they were communicated in parallel.

The Xyce parser performs two passes to do its work.  The first pass is
mostly in serial on proc 0, and the second pass is mostly parallel.
Device lines are not processed in the first pass other than
tokenization.  In the second pass, the tokenized device lines are first
communicated in parallel, and then fully processed/resolved.  This
includes all the parameters found on those device lines.

Subcircuits are a special case.  In some aspects, the parser treats
subcircuit "X" lines like another device line, and in other respects
treats them differently.  The biggest differences is that the subcircuit
"X" line is resolved during the first pass.  Note, the fully expanded
subcircuit is resolved in the second pass.  Just the "X" line is
resolved in the first pass.

Note, this choice, to treat "X" lines kind of like device lines, but
also kind of not, made the code pretty confusing. But I digress.

Unfortunately, resolving X lines on proc 0 during pass 1  meant
that parameters on the "X" line were resolved prior to parallel
communication, and some important attributes of these resolved params
were lost in the communication process, without any way to recover
them.

Specifically, any external dependencies of the AST on .param and/or
.global_params went away.  Also, any information relating to local
variation went away as well.  EXPR-based parameters are communicated
(via the Param::pack and Param::unpack functions) as strings.
So, any information that can't be captured by the string is lost.

This meant that subcirucit parameters couldn't have any complex
dependencies or attributes.  Basically, they could ONLY be constants.
If they were constants, or very simple expressions, then they
could converted to numbers and/or strings prior to communication,
and there was sufficient information to use them correctly after
communication.

This fix involves changing the order of operations.  Subcircuit "X" lines
are still *mostly* resolved in the first pass, but the X-parameters are not.
The X-parameters are only fully resolved *after* parallel communication.
This makes it possible for all the needed attributes to be generated
correctly, locally on-processor during pass 2.

Note, the meaninful changes in this commit are mostly in the
distribution tools (N_IO_DistToolDefault, N_IO_DistToolBase and
N_IO_DistToolDevBalance).  A bunch of other files are included
in this commit b/c I updated a lot of debug output, mostly to
switch from std::cout to Xyce::dout, which works much better
in parallel.

Most of my work centered around the N_IO_DistToolDefault class, but I
have implemented the same changes int the N_IO_DistToolDevBalance class.
While doing this, I realized that some of my prior work on local
varation had not been implemented in the dev-balance class, so I
took care of that in this commit as well.  This required moving one
function from the N_IO_DistToolDefault class to the DistToolBase
class, among other changes.

Note that one of the distribution tools, N_IO_DistToolFlatRoundRobin, is
not touched in this commit, because it doesn't handle subcircuits.