Remove temperature interpolation from UG and RG
This is related to this issue:
https://cee-gitlab.sandia.gov/Xyce/backlog/project-backlog/-/issues/649
"Remove the temperature interpolation methods"

These features have been slated for removal for (at this point) nearly 5
years.  There have been warnings in the release notes for the past 10
releases about it.

I think it is time to finally actually remove it.  The first step is
removing the documentation, which is the point of this commit.