Slightly upadte subcircuit parameter documentation
This was prompted by my work on:

https://cee-gitlab.sandia.gov/Xyce/backlog/project-backlog/-/issues/609
"Global parameters that are applied to subcircuit instance parameters don't work in parallel"

I was mostly concerned with removing text that contradicted the new
feature enabled by that work.  None of the text was quite like that, but
the text was out of date with respect to the PARAMS: keyword, which is
now optional.  So, I updated that text.