Don't use automated dependency check for windows install
CMake can look for dependencies that an executable needs
and package those with the executable in the installer.
Under linux and mac os this does not include system libraries,
but under Windows it appears that it does find system objects.
The simple path forward is to not use dependency checking
on Xyce.exe

https://cee-gitlab.sandia.gov/Xyce/backlog/project-backlog/-/issues/212