Fixes sequential source stepping for corrected transient sources
Recently sources with only transient specifications have been fixed
to use the correct value in simultaneous source stepping.  This commit
serves to fix the logic such that the DC source value obtained for
sequential source stepping is now correct.  This required moving the
call to computeF above the call to the device package to get the source
devices, their DC values, and their connectivity.  If the DC value and
connectivity is not correct, then the logic in the sequential source
stepping code will not scale all the non-zero DC sources in ascending
order of connectivity.  Furthermore, the method in the device manager
for obtaining the sources, their DC values, and their connectivity
needed to be modified to update dependent parameters.  Otherwise, the
DC value could be wrong if parameters are used in the source description.

https://cee-gitlab.sandia.gov/Xyce/backlog/project-backlog/-/issues/651