Remove use of "RUNTIME_DEPENDENCIES" in Xyce INSTALL command
The RUNTIME_DEPENDENCIES parameter in the CMake install command
for Xyce should pick up non-system supplied, external shared
libraries that Xyce uses and package them as part of the Xyce
installer.  Under MacOS it sort of works but fails to correctly
set the @rpath@ on the libraries.  Under Linux is can generate
an error if there are multiple copies of a dependent library
found and under Windows it picks up system libraries that it
claims it will not.  Thus, we will not use the
RUNTIME_DEPENDENCIES option and just manually pack extra,
3rd party libraries like we do with the Intel Math libraries.

https://cee-gitlab.sandia.gov/Xyce/backlog/project-backlog/-/issues/212