Correct release variable in cmake
When I merged the release branch, the changes to configure.ac got
flagged as a conflict that I had to resolve, but the changes I had
made to cmake/features_modes.cmake did not.  That would have left the
build claiming it was a release rather than a development branch
build.