Add interpolation and extrapolation to S-parameter model
The commit adds interpolation and extrapolation methods
to the S-parameter model. Two interpolation methods, linear and a
special form of spline, are supported. Both the passive part and
Isc use the same interpolation and extrapolation methods. The interpolation
and extrapolation of the complex data use real and imaginary based
interpolation. The interpolation method can be selected by setting
model parameter interpolation. The default is linear interpolation method

https://cee-gitlab.sandia.gov/Xyce/backlog/project-backlog/-/issues/46