Make PSP103 DTA parameter be instance/model
The PSP 103.4 model in Xyce was written by its authors to set the
DTA (temperature difference from ambiant) as a model parameter.  This
not only makes no sense, it has been requested by users that it be
instance instead.

This commit patches the PSP103 Verilog-A to accomplish this.  Two bits
are needed:
   - add attributes to the DTA parameter declaration to make Xyce/ADMS
     create both an instance and model version
   - move code that initializes some variables using DTA out of the
     @(initial_model) block and into @(initial_instance)

A companion test case will be created to go with this change.

I am also updating the reference guide tables impacted by this change
in this commit.