Fix voltage limiting in BSIMSOI3 device
The voltage limiting in nonlinear devices is a method to help
convergence. This commit fixes a mistake in the voltage limiting
related calculation in BSIMSOI3 which led to convergence issues in CMOS7 circuits.
The voltage limiting related calculation involves calculating difference
between voltage drops before and after voltage limiting.
However, Vps_orig which is the original voltage drop before voltage
limiting is not updated correctly when vds >= 0. This led to incorrect
voltage limiting terms related to the internal and external
body nodes.

In the customer circuit, this led to different behavior
when using gminstepping alone and using gminstepping as part of
Spice strategy. One led to convergence and the other led to failure for
DC op calculation.

This issue also affects simulation when voltage limiting is turned off
for bsimsoi3 device. The initial junction voltage is applied
for the first newton iteration. The voltage limiting related calculation
is done in this case. The commit fixes the voltage limiting issue in
bsimsoi3 device

https://cee-gitlab.sandia.gov/Xyce/backlog/project-backlog/-/issues/719