Fix error in limiter detection
This is a fix for this issue:
https://cee-gitlab.sandia.gov/Xyce/code/uur-proprietary/Xyce/-/issues/469
"Fix problem with Xyce/ADMS code generation when $limit arguments have
expressions with function calls"

The test in xyceBasicTemplates' "collectLimiters" template only
recognized limiting assignments where the `$limit` function is the
only function call on the RHS, and would be confused by any RHS that
had additional function calls in the `$limit` arguments.

This commit fixes it so that all assignments that `$limit` in their
list of RHS functions are recognized.

Once that was fixed, it turned out there was a problem in code
generation of limiter assignments.  The limiterAssignment was
gleefully emitting code that presumed function precomputation, but was
not doing function precomputation.  The simplest fix for that was just
to disable function precomputation for those assignments as it is of
no benefit.  (Function precomputation is a thing that really helps out
when generating code for assignments that need differentiation, but
limiting assignments do not need that so it's a waste of time to
retrofit precomputation onto them.)