Adds new setBreakPoints to the StepErrorControl class
The .OPTIONS OUTPUT can let users set a list of breakpoints that are
user-defined, for output, or for adjoint calculations.  A user employed
`outputtimepoints` to set a list of breakpoints for training a surrogate
model.  This list had 100,000 time points in it and Xyce took a long time
to set up the simulation before executing the transient analysis.  That is
because the process for merging this list into the set of breakpoints was
done one-at-a-time, which does not scale for very large lists.

A new setBreakPoints method is being added to the StepErrorControl class
that will process the whole list of breakpoints at once, merging it in
with the current list of breakpoints.  This approach scales and substantially
improves performance when long lists are provided to .OPTIONS OUTPUT

https://cee-gitlab.sandia.gov/Xyce/backlog/project-backlog/-/issues/740