Small sensitivity analysis bugfix
This is a fix for this issue:
https://cee-gitlab.sandia.gov/Xyce/backlog/project-backlog/-/issues/729

"Expression-based objective functions for .SENS conflate I(VA) with
V(VA) (nodes and devices w/ same names)"

This was reported by an external user.  The "group" class used by
sensitivity analysis objective function expressions was not smart enough
to tell the difference between I(VA) and V(VA), when the net list
contained a voltage node named VA and also a device by that name.

This was caused by my taking a shortcut in the group class, and was easy
to fix.   Some of the objective management code (in
N_NLS_ObjectiveFunctions.h) also needed a small rewrite.