Fixes MOR computation to work in parallel
The model-order reduction analysis in Xyce was not written to be
correct in parallel.  This work leverages the block linear algebra
that is used in HB, so it is only a limitation of the implementation.
Just a few changes to the assembly of the block linear algebra makes
this work.  There are some additional issues with the sparsification
and autodetection algorithms.  Those will be addressed in a separate
issue.

https://cee-gitlab.sandia.gov/Xyce/backlog/project-backlog/-/issues/751