Add initial BSIM CMG version 111.2.1 files
This commit is for this issue:
https://cee-gitlab.sandia.gov/Xyce/backlog/project-backlog/-/issues/753
"Implement version 111.2.1 of the BSIM CMG device"

This commit adds the BSIM CMG version 111.2.1 verilog-A files, completely
unmodified.  They will not yet compile for Xyce/ADMS, and will require
modifications to make it possible.

In addition to the BSIM CMG files, I have created a makefile by mimicing what
Tom Russo did for the version 110 device.   I have also created a README_Xyce
file to document the changes.  Currently that file is empty except for the
header.