Change construction of graph objects to be static
The Epetra package in trilinos is being deprecated. This commit is
part of the early stages in refactoring references to Epetra objects
in Xyce (part of the groundwork to ultimately replace Epetra with
Tpetra). Instead of dynamically building the graph, the allocation is
now done statically, and is moved into the base Graph abstraction
layer. This change reduces the number of implementation dependent
operations.

Issue:
https://cee-gitlab.sandia.gov/Xyce/backlog/project-backlog/-/issues/727
Task:
https://cee-gitlab.sandia.gov/Xyce/backlog/project-backlog/-/issues/731