Fix the `lexp macro to remove ADMS lexical error
This macro uses a ternary operator, and it seems that in order for ADMS
to parse it properly, it needed more parenthesis.  I've encountered
similar problems parsing ternaries in other contexts, so this probably
isn't surprising.

I have also added information about the last two commits to the
README_Xyce file.

This commit is for this issue:
https://cee-gitlab.sandia.gov/Xyce/backlog/project-backlog/-/issues/753
"Implement version 111.2.1 of the BSIM CMG device"