Add @initial_instance statement
This commit is for this issue:
https://cee-gitlab.sandia.gov/Xyce/backlog/project-backlog/-/issues/753
"Implement version 111.2.1 of the BSIM CMG device"

In order to work with ADMS, a bunch of the initial setup code needed to
be placed inside of an @initial_instance statement, followed by a
"begin" and a subsequent "end" many lines later.

This is in part to save on runtime, and also to allow some of these
variables to be used on node-collapse conditionals.