Convert params to be both instance and model
This commit is for this issue:
https://cee-gitlab.sandia.gov/Xyce/backlog/project-backlog/-/issues/753
"Implement version 111.2.1 of the BSIM CMG device"

At the top of the bsimcmg_parameters.include file, there is a block of
parameters identified as being both "model" and "instance" parameters.
In order for ADMS/Xyce to handle these correctly, it is necessary to add
a special attribute, xyceAlsoModel="yes" to their declarations.

These are declared using macros from the bsimcmg_macros.include file, so
to do this properly required new macros to be created.  Following the
established pattern from the level=110, the new macros all have the
suffix "_BOTH" in their name.

With this change, ADMS can now compile the model.

The README_Xyce file has been updated to document these changes, as well
as other recent changes such as the @initial_instance modification.