Removes dynamic construction of matrix object in MOR code
The only use of dynamic matrix construction was in the model-order
reduction code.  That has been replaced with the use of the new
method for creating static graphs and standard methods for setting
matrix values.

This commit also removes some direct creation and use of Epetra
objects in MOR, as well as direct solvers from Amesos.  More of this
code needs to be rewritten to enable MOR to work after the Tpetra
transition.

Issue:
    https://cee-gitlab.sandia.gov/Xyce/backlog/project-backlog/-/issues/727
Task:
    https://cee-gitlab.sandia.gov/Xyce/backlog/project-backlog/-/issues/732