Update the ADMS back-end to handle $error keyword
This is based on advice from Tom Russo.

The BSIM-CMG device, version 111 uses the $error keyword to report fatal
device errors.   The back-end of ADMS/Xyce did not support this keyword,
so it resulted in uncompilable C++ code.

This commit is for this issue:
https://cee-gitlab.sandia.gov/Xyce/backlog/project-backlog/-/issues/753
"Implement version 111.2.1 of the BSIM CMG device"