Fix an issue in ADMS/Xyce back-end dependencies
This came up in my work on the BSIM-CMG version 111.
So this commit is for this issue:
https://cee-gitlab.sandia.gov/Xyce/backlog/project-backlog/-/issues/753
"Implement version 111.2.1 of the BSIM CMG device"

The ADMS back end was a little sloppy in how it handled ternary
operators.  The front end lexer has had difficulty as well, but
that isn't the focus of this commit.

This fix was found quickly with Tom Russo's guidance.  Here is his
description.

There was a bug in the output of ternary operator derivatives if
both the value-when-true and value-when-false (in the derivative) were
identically zero.  The ternary derivative should then have been
identically zero, but instead it was incorrectly being emitted
as (condition)?0.0:0.0.

This means that any expression in which a zero-derivative ternary was
a subexpression was not getting properly reduced, and this led to some
weird issues in the BSIM-CMG 111.  Now, if both value-if-true and
value-if-false of the derivative are identically zero, then the
derivative of the ternary is returned as "0.0".  This allows
downstream processing to emit more efficient code (like a-b gets
emitted as "(-b)" if a is 0.0, or a*b is 0.0 if either a or b is 0.0).