Add binning min/max vars to BSIM-CMG 111
This commit is for this issue:
https://cee-gitlab.sandia.gov/Xyce/backlog/project-backlog/-/issues/753
"Implement version 111.2.1 of the BSIM CMG device"

Model binning works for the BSIM-CMG , and while it works just fine
whether or not LMIN/LMAX and NFINMIN/NFINMAX are defined by the device
code, Xyce will spew warnings about unrecognized model parameters
unless the dummy parameters are defined in the device.  This commit
adds them to the BSIM-CMG level 111.

The output variables defined by the original BSIM-CMG 111 authors are
already automatically set up.

This commit also adds several user-requested output variables.
They include "VDSAT", VDS, VGS, and VBS output variables.
None of these variables are in the set defined by the
original source code.  VDSAT is in fact a duplicate of what the
authors called "VDSSAT", and the others are just voltage drops across
the external nodes.  These variables have been added back in to the
Verilog-A source.

With respect to both of these issues (output variables and binning
variables) the other BSIM-CMG model levels previously were modified
in this manner.