Add M= multiplier to BSIMCMG 111
This commit is for this issue:
https://cee-gitlab.sandia.gov/Xyce/backlog/project-backlog/-/issues/753
"Implement version 111.2.1 of the BSIM CMG device"

The original version of BSIM CMG 111 device didn't have the "M"
multiplier set up.  This was also the case for previous versions of the
BSIMCMG, although those versions had *some* artifacts related to M, just
not enough.

Most simulators implement M= multipliers at the simulator level, not the
model level.  This is the right thing to do, but we haven't done that in
Xyce yet.  So, for now we have to hack devices to support it.

To set up the multiplier in the 111 device, I followed the pattern
established by Tom the 110.