Final removal of direct Epetra usage in MOR analysis
The Epetra-dependent utilities that bridge the computation between
the original system and projected system have been moved to helper functions
in the linear algebra package.  This incapsulation has improved the conversion
from Epetra to Tpetra by eliminating the direct use of Epetra in analysis classes.
Much of the MOR analysis could be performed using the abstract linear algebra
primatives, so parts of the code have also been refactored to enable this.

Issue:
    https://cee-gitlab.sandia.gov/Xyce/backlog/project-backlog/-/issues/727
Task:
    https://cee-gitlab.sandia.gov/Xyce/backlog/project-backlog/-/issues/761