Enable gmin in b3soi device to work consistently with other devices
The gmin in b3soi is much smaller than the gmin used in other
nonlinear devices. The commit makes it possiable to turn off
the extra scaling factor of gmin in b3soi device. The device
option b3soigminscaling is used to turn on or off the scaling.
The gmin scaling is on  by default in Xyce.

https://cee-gitlab.sandia.gov/Xyce/backlog/project-backlog/-/issues/721