Add the BISM CMG 111 to the device package
This commit is for this issue:
https://cee-gitlab.sandia.gov/Xyce/backlog/project-backlog/-/issues/753
"Implement version 111.2.1 of the BSIM CMG device"

This work isn't quite finished yet.  However, I am at a point in
my development of the BSIM CMG 111 device that I am now building the
device directly in the Xyce source tree.  Prior to this point I was
linking it in dynamically.

This will get refined every time I make corrections to the verilog-A
code for this device.