Fix noise analysis for correct flicker noise
This commit was prompted by my work on this issue:
https://cee-gitlab.sandia.gov/Xyce/backlog/project-backlog/-/issues/753
"Implement version 111.2.1 of the BSIM CMG device"

The flicker noise model in version 111 of the BSIMCMG contains a
flicker noise bugfix.  The argument to flicker_noise now has a sign
variable, "sigvds", which is +1 or -1 depending on the sign of Vds.

This change was based on the guidance in this 2020 paper by
Geoffrey Coram:
"Flicker Noise Formulations in Compact Models"
in IEEE Transactions on CAD
DOI: 10.1109/TCAD.2020.2966444

The BSIM CMG model forced me to make this change. However, it
is a more general issue that was bound to impact us sooner
or later, as the paper recommends that all flicker noise models
(in any device model) going forward contain this sign correction.

Small signal noise (the only kind Xyce currently has) everything is
based around RMS values.  So the computed noise densities should
be positive.  Howevr, the way Xyce noise analysis was written,
the always-positive condition on the inputs was not enforced by
the analysis code.  It was just implicitly assumed.  So, when the
BSIMCMG added the correct (sometimes negative) sign term, that
caused Xyce's small signal noise outputs to go negative
incorrectly.

This commit changes the function NOISE::solveAdjointNOISE_ function so
that the contributions from all the noise sources are subject to
fabs().   This wouldn't be correct for large-signal noise, but works
for small-signal.