Optimize model by expanding initial_instance
This commit is for this issue:
https://cee-gitlab.sandia.gov/Xyce/backlog/project-backlog/-/issues/753
"Implement version 111.2.1 of the BSIM CMG device"

For variables that should only be computed once during setup, and not
repeatedly during the solve, one can put such calculations inside the
"intial_instance" block.  For Xyce/ADMS, all calculations inside this
block will be in the processParams functions, rather than
updateIntermediateVars.

I previously set this up, but was cautious, so some things were left out
of the initial_instance block that should have been inside of it.  This
commit corrects this issue.