Optimize BSIM-CMG 111 without self-heating
This commit is for this issue:
https://cee-gitlab.sandia.gov/Xyce/backlog/project-backlog/-/issues/753
"Implement version 111.2.1 of the BSIM CMG device"

This commit is similar to one that was performed on the 110 model.
Quoting from the main commit for that work:

The BSIM-CMG has a largish block of code to set temperature-dependent
variables.  Because the local temperature rise is a solution variable,
this cannot possibly be set up to be a "bias-independent" calculation by
ADMS, and cannot be placed in the processParams function where
bias-independent stuff goes.  These computations *must* be in
updateIntermediateVars.

But when self-heating is off, these computations are wasteful, as
temperature is not changing.

This commit massages the Verilog-A in such a way that it is possible to
wrap the block of temperature-dependent code in a conditional that
checks to see if the temperature has changed.  If it hasn't, it'll just
reuse previously computed values for these variables.

On the 110 model, Tom observed as much as a 30% speed improvement on the
"inverter_transient" test case.  I have not attempted to reproduce
that study for the 111, but I believe that this optimzation will
produce a similar speedup in that device.

More discussion about this can be found in the old bugzilla issue for the 110:
http://joseki-vm.sandia.gov/bugzilla/show_bug.cgi?id=894

The details of the 111 implementation are slightly different than for
the 110, but the conceptual idea is the same.  In the 111, there exists
a file, bsimcmg_initialization.include, which is the natural place to
put the temperature-variable initializations.  This file didn't exist in
the 110.