Fix temperature equ optization in BSIMCMG 111
This commit is for this issue:
https://cee-gitlab.sandia.gov/Xyce/backlog/project-backlog/-/issues/753
"Implement version 111.2.1 of the BSIM CMG device"

My first cut at optimizing temperature equations, so that they are only
evaluated when the temperature changes had some mistakes.

The tricky part about doing an optimization like this is that :

(1) it puts all the temperature equations inside of a conditional and

(2) any variables that are *implicitly* declared inside that conditional become
local variales.  So, care must be taken to initialize them outside of
that conditional beforehand, so they are not local.  But, because
verilog doesn't require that they be explicitly declared, it is hard to
track down which variables have been implicitly declared inside the
conditional block.