Fix smoothing macros to avoid NaNs in derivatives
This commit is for this issue:
https://cee-gitlab.sandia.gov/Xyce/backlog/project-backlog/-/issues/753
"Implement version 111.2.1 of the BSIM CMG device"

When self-heating is turned on in this model, it becomes necessary for
the temperature equations to, in addition to computing values, also
compute derivatives for the Jacobian matrix.

The problem is that the temperature equations use smoothing function
macros.  They are called unconditionally.  These smoothing macros
include sqrt functions.  When differentiated, you get a 1/sqrt term.
The problem comes when all the arguments to the smoothing
function are zero.  Most (probably all, but I haven't exhaustively
checked) CMC tests use default parameters for
these smoothing functions, which are zero, so this can happen.

So, this leads (roughly) to a 1/sqrt(0) term, which is Inf, and in the
next equation causes an NaN in a matrix derivative.

I've fixed these macros to take advantage of ternary operators.  Now,
if all the arguments are zero, they just return zero.  If not, they
evaluate the original function.  The ADMS/Xyce back-end is smart enough
to not compute a derivative in this case.

With this fix, a bunch of cryomod=1 tests that were failing now pass.