Modifies graph assembly to be static, instead of dynamic
In the transition from Epetra to Tpetra, removing dynamic graph assembly
will be necessary.  This commit removes it from the HB block jacobi preconditioner,
which is still written directly in Epetra.  It is difficult to rewrite this
preconditioner using the abstract interface, so writting a Tpetra version
of it will have to wait until later in this transition process.

    Issue:
    https://cee-gitlab.sandia.gov/Xyce/backlog/project-backlog/-/issues/727
    Task:
    https://cee-gitlab.sandia.gov/Xyce/backlog/project-backlog/-/issues/731