Remove unused separated solution graph and map fnc
This commit removes a set of graph builder helper functions. This
particular group of functions would separate the solution variables
related to linear and nonlinear devices. The setup of the separated
solution variable objects utilizes a graph api that is closely coupled
to Epetra. In order to transition Xyce from Epetra to Tpetra, these
functions would need to be refactored.

However, the separated graph fuctions and objects are currently not in
use. The only code that used the separated graph helper functions was
refactored a few years ago, but the helper functions remained for future
use. Given that there is still no immediate plan to use these functions,
and there are no active tests for them either, the decision was made to
remove them.

Issue:
Xyce/backlog/project-backlog#727
Task:
Xyce/backlog/project-backlog#731