Fix for CSV noise output
This is for this issue:
https://cee-gitlab.sandia.gov/Xyce/backlog/project-backlog/-/issues/613
"CSV output format should (maybe optionally) put quotes around variable/column names"

It was pointed out by an external user that the CSV output for noise had
a flaw in it.  The flaw is that many noise outputs are quantities like
DNO(device,submodel), where there is a comma in the name.   Since the
variable fields are separated by commas, this confuses most applications
that would read in this file.

This commit implements two solutions.  They are:

1.  If the noise output format is CSV, and no alternative delimiter is
    specified, then put quotes around all the variable names in the
    header.  So, instead of:

Index, ONOISE, DNO(device,model)

instead have
"Index", "ONOISE", "DNO(device,model)"

2.  Allow the user to specify the DELIMITER on the .PRINT line, and
    expand to allow for more delimiters including colon and semicolon,
    and also allow this DELIMITER parameter to apply to CSV output
    (noise or otherwise).  The DELIMITER parameter on the .PRINT line
    was already supported, but it only applied to *prn (std) output,
    and it only supported tabs and commas before this commit.   Now,
    it also applies to *CSV format.  This kind of means that *prn
    and *csv are basically the same format, except with different
    default delimiters.

    In addition to TAB, COMMA, SEMICOLON and COLON, the user can also
    specify a single-character string for the delimiter.  So for
    example, DELIMITER=";" is allowed as well.