Documentation of expanded PRINT DELIMITER param
This is for this issue:
https://cee-gitlab.sandia.gov/Xyce/backlog/project-backlog/-/issues/613
"CSV output format should (maybe optionally) put quotes around variable/column names"

This is now documented in the user guide and reference guide.

Also included in this commit is a small change to the delimiter-relevant
code in the output manager class.  I had set up the new delimiters to
allow DELIMITER="c" on the netlist .PRINT line, where "c" was only
allowed to be a quoted single ASCII character.  I had not realized that
the command line option "-delim" already support strings as delimiters,
with no limit on the size of the string.  So, I updated the code to
allow delimiters that are multi-character strings.  They still must be
quoted to differentiate from delimiters like COMMA.

So, now (for example) this is allowed:  .PRINT DC DELIMITER="_fred_"

Not sure *why* someone would use this, but they could if they wanted to.