Fix for numerical sensitivities with option scale
This commit is related to:

"Error in MOSFET numerical sensitivities, when using .option SCALE"
https://cee-gitlab.sandia.gov/Xyce/backlog/project-backlog/-/issues/694

This was reported by an external user.   When attempting to use
sensitivity analysis with the BSIM4, the code exited with an error.

The BSIM4, unfortunately, doesn't have analytical derivatives
implemented for sensitivity analysis.  So, for the portion of
the sensitivity calculation for which the device model is responsible,
Xyce takes a numerical derivative.

The device-level numerical derivative code is generic, and it was not
instrumented to understand parameters that have been
scaled via the ".option scale" command.  In the use case reported by the
external users, the netlist was applying an "option scale" command, and
the sensitivity parameters were L,W on the MOSFET instance line.

The details aren't important, but basically L,W got scaled every time
the derivative was computed, and were not properly un-scaled.  This
messed up the entire calculation.

This commit fixes the issue.  It fixes it for both numerical vector
(used by DC and tran) and matrix (used by AC) sensitivities in
the instance class.