Bugfix for the AC sensitivities group
This is related to this issue:
"Expression-based objective functions for .SENS conflate I(VA) with V(VA) (nodes and devices w/ same names)"
https://cee-gitlab.sandia.gov/Xyce/backlog/project-backlog/-/issues/729

I previously commited code changes to fix this issue.  However, in doing
so I unintentionally broke AC sensitivities.

The breakage to AC sensitivities happened in part because I forgot that
AC sensitivities have their own special group class.  This is easy to
forget, because unlike the other group classes it sits inside the
N_ANP_AC.h and N_ANP_AC.C files, which are in the AnalysisPKG directory.
The other group classes all sit in the same directory as the expression
classes, in UtilityPKG/ExpressionSrc.

AC sensitivities need their own class because AC analysis uses its own
special set of vectors to perform the analysis.  It does this to support
complex vectors and matrices via a real equivalent form.  The main group
classes, which sit by the expression library don't know about these
specialized objects.  So, if you try to use the main group class with AC
analysis, it doesn't work.  It points to the wrong (non-complex)
solution vector.

In my modifications for this issue, I added new derived functions to
the mainExpressionGroup object, for obtaining current variables.  I
needed to also add an overloaded function for current variables to the
AC group, but negleected to do so.    As a result, the AC sensitivity
analysis expressions were incorrectly calling the new main group
function.

This is fixed now.